+++ crates/bevy_ui/src/render/ui_texture_slicer.wgsl
