+++ crates/bevy_ui/src/render/ui_texture_slicer_pipeline.rs
